[REFACTOR] #267 닉네임 팝오버 상태관리 postdetailpage에서 postdetailinfopanel로 이동
